tech?(api): move service getMasteryPercentage into evaluation bounded context
